Refactoring directed graph code from nutrient.
