+++ ApsimNG/Presenters/DirectedGraphPresenter.cs
@@ -1,10 +1,16 @@
-﻿namespace UserInterface.Presenters
-{
-    using Models.Interfaces;
-    using System.Drawing;
-    using System.IO;
-    using Views;
+﻿using APSIM.Shared.Graphing;
+using APSIM.Shared.Utilities;
+using Models.Core;
+using Models.Interfaces;
+using Models.Soils;
+using Models.Soils.Nutrients;
+using System.Drawing;
+using System.IO;
+using System.Linq;
+using UserInterface.Views;
 
+namespace UserInterface.Presenters
+{
     /// <summary>
     /// This presenter connects an instance of a Model with a 
     /// UserInterface.Views.DrawingView
@@ -57,8 +63,82 @@
         /// <summary>Populate the view object</summary>
         private void PopulateView()
         {
+            CalculateDirectedGraph();
             if (model.DirectedGraphInfo != null)
                 view.DirectedGraph = model.DirectedGraphInfo;
+        }
+
+        /// <summary>Calculate / create a directed graph from model</summary>
+        public void CalculateDirectedGraph()
+        {
+            DirectedGraph oldGraph = model.DirectedGraphInfo;
+            if (model.DirectedGraphInfo == null)
+                model.DirectedGraphInfo = new DirectedGraph();
+
+            model.DirectedGraphInfo.Begin();
+
+            bool needAtmosphereNode = false;
+
+            IModel nutrient = model as IModel;
+            foreach (NutrientPool pool in nutrient.FindAllInScope<NutrientPool>())
+            {
+                Point location = default(Point);
+                Node oldNode;
+                if (oldGraph != null && pool.Name != null && (oldNode = oldGraph.Nodes.Find(f => f.Name == pool.Name)) != null)
+                    location = oldNode.Location;
+                model.DirectedGraphInfo.AddNode(pool.Name, ColourUtilities.ChooseColour(3), Color.Black, location);
+
+                foreach (CarbonFlow cFlow in pool.FindAllChildren<CarbonFlow>())
+                {
+                    foreach (string destinationName in cFlow.DestinationNames)
+                    {
+                        string destName = destinationName;
+                        if (destName == null)
+                        {
+                            destName = "Atmosphere";
+                            needAtmosphereNode = true;
+                        }
+
+                        location = default(Point);
+                        Arc oldArc;
+                        if (oldGraph != null && pool.Name != null && (oldArc = oldGraph.Arcs.Find(f => f.SourceName == pool.Name && f.DestinationName == destName)) != null)
+                            location = oldArc.Location;
+
+                        model.DirectedGraphInfo.AddArc(null, pool.Name, destName, Color.Black, location);
+
+                    }
+                }
+            }
+
+            foreach (Solute solute in nutrient.FindAllInScope<ISolute>())
+            {
+                Point location = new Point(0, 0);
+                Node oldNode;
+                if (oldGraph != null && solute.Name != null && (oldNode = oldGraph.Nodes.Find(f => f.Name == solute.Name)) != null)
+                    location = oldNode.Location;
+                model.DirectedGraphInfo.AddNode(solute.Name, ColourUtilities.ChooseColour(2), Color.Black, location);
+                foreach (NFlow nitrogenFlow in nutrient.FindAllChildren<NFlow>().Where(flow => flow.sourceName == solute.Name))
+                {
+                    string destName = nitrogenFlow.destinationName;
+                    if (destName == null)
+                    {
+                        destName = "Atmosphere";
+                        needAtmosphereNode = true;
+                    }
+                    location = default(Point);
+                    Arc oldArc;
+                    if (oldGraph != null && solute.Name != null && (oldArc = oldGraph.Arcs.Find(f => f.SourceName == solute.Name && f.DestinationName == destName)) != null)
+                        location = oldArc.Location;
+
+                    model.DirectedGraphInfo.AddArc(null, nitrogenFlow.sourceName, destName, Color.Black, location);
+                }
+            }
+
+            if (needAtmosphereNode)
+                model.DirectedGraphInfo.AddTransparentNode("Atmosphere");
+
+
+            model.DirectedGraphInfo.End();
         }
 
     }
